Reduce the number of files loaded at "require 'pg'"
And use autoload to load optional classes when they are used.
