+++ lib/pg/binary_encoder/timestamp.rb
