--- lib/pg/constants.rb
@@ -1,12 +1,0 @@
-# -*- ruby -*-
-# frozen_string_literal: true
-
-require 'pg' unless defined?( PG )
-
-
-module PG::Constants
-
-	# Most of these are defined in the extension.
-
-end # module PG::Constants
-
