+++ lib/pg/text_decoder/timestamp.rb
@@ -1,27 +1,8 @@
 # -*- ruby -*-
 # frozen_string_literal: true
 
-require 'date'
-require 'json'
-
 module PG
 	module TextDecoder
-		class Date < SimpleDecoder
-			def decode(string, tuple=nil, field=nil)
-				if string =~ /\A(\d{4})-(\d\d)-(\d\d)\z/
-					::Date.new $1.to_i, $2.to_i, $3.to_i
-				else
-					string
-				end
-			end
-		end
-
-		class JSON < SimpleDecoder
-			def decode(string, tuple=nil, field=nil)
-				::JSON.parse(string, quirks_mode: true)
-			end
-		end
-
 		# Convenience classes for timezone options
 		class TimestampUtc < Timestamp
 			def initialize(**kwargs)
