+++ lib/pg/text_encoder/json.rb
@@ -1,0 +1,14 @@
+# -*- ruby -*-
+# frozen_string_literal: true
+
+require 'json'
+
+module PG
+	module TextEncoder
+		class JSON < SimpleEncoder
+			def encode(value)
+				::JSON.generate(value, quirks_mode: true)
+			end
+		end
+	end
+end # module PG
